Move content from GradesActivity to GradesFragment
