+++ app/src/main/res/layout-land/activity_grades.xml
@@ -34,7 +34,12 @@
 
         </com.google.android.material.appbar.AppBarLayout>
 
-        <include
+        <FrameLayout
+            android:id="@+id/contentFrame"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" />
+
+        <!--<include
             layout="@layout/layout_all_errors"
             android:layout_width="fill_parent"
             android:layout_height="fill_parent" />
@@ -129,7 +134,7 @@
                 app:backgroundTint="@color/color_primary"
                 app:cornerRadius="@dimen/material_corner_radius" />
 
-        </LinearLayout>
+        </LinearLayout>-->
 
     </LinearLayout>
