+++ app/src/main/res/layout/activity_grades.xml
@@ -1,109 +1,13 @@
 <androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
 
-    <LinearLayout
-        android:id="@+id/gradesLayout"
+    <FrameLayout
+        android:id="@+id/contentFrame"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <com.google.android.material.appbar.AppBarLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <androidx.appcompat.widget.Toolbar
-                android:id="@+id/main_toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:background="@android:color/white"
-                android:elevation="12dp"
-                android:theme="@style/TumToolbar"
-                app:popupTheme="@style/TumToolbarPopUp">
-
-                <Spinner
-                    android:id="@+id/filterSpinner"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content" />
-
-            </androidx.appcompat.widget.Toolbar>
-
-        </com.google.android.material.appbar.AppBarLayout>
-
-        <include
-            layout="@layout/layout_all_errors"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-        <FrameLayout
-            android:id="@+id/contentContainer"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1">
-
-            <FrameLayout
-                android:id="@+id/chartsContainer"
-                android:layout_width="match_parent"
-                android:layout_height="200dp"
-                android:padding="@dimen/material_default_padding"
-                android:visibility="invisible">
-
-                <com.github.mikephil.charting.charts.PieChart
-                    android:id="@+id/pieChartView"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:visibility="visible" />
-
-                <com.github.mikephil.charting.charts.BarChart
-                    android:id="@+id/barChartView"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:visibility="gone" />
-
-            </FrameLayout>
-
-            <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-                android:id="@+id/swipeRefreshLayout"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-
-                <se.emilsjolander.stickylistheaders.StickyListHeadersListView
-                    android:id="@+id/gradesListView"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:clipToPadding="false"
-                    android:divider="@color/list_divider"
-                    android:paddingTop="200dp"
-                    android:scrollbars="vertical" />
-
-            </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-        </FrameLayout>
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="#EEE"
-            android:padding="@dimen/material_default_padding">
-
-            <TextView
-                android:id="@+id/averageGradeTextView"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:fontFamily="sans-serif-medium"
-                android:textAppearance="@style/TextAppearance.AppCompat.Medium"
-                android:textColor="#757575"
-                android:textSize="@dimen/semilarge_text_size"
-                android:visibility="visible"
-                tools:text="Average grade: 1.0" />
-
-        </FrameLayout>
-
-    </LinearLayout>
+        android:layout_height="match_parent" />
 
     <!-- The navigation drawer -->
     <include layout="@layout/navigation_drawer" />
